+++ packages/observable-client/src/chainHead/streams/follow.ts
@@ -1,45 +1,89 @@
+import { BlockHeader, blockHeader } from "@polkadot-api/substrate-bindings"
 import {
   ChainHead,
   FollowEventWithRuntime,
   FollowResponse,
   StopError,
 } from "@polkadot-api/substrate-client"
-import { Observable, Subscription, noop, share } from "rxjs"
+import {
+  Observable,
+  ObservedValueOf,
+  Subscription,
+  concatMap,
+  connectable,
+  noop,
+  of,
+  share,
+} from "rxjs"
+
+const withInitializedNumber =
+  (getHeader: (hash: string) => Promise<BlockHeader>) =>
+  (source$: Observable<FollowEventWithRuntime>) =>
+    source$.pipe(
+      concatMap((event) => {
+        return event.type !== "initialized"
+          ? of(event)
+          : getHeader(event.finalizedBlockHashes[0]).then((header) => ({
+              ...event,
+              number: header.number,
+              parentHash: header.parentHash,
+            }))
+      }),
+    )
 
 export const getFollow$ = (chainHead: ChainHead) => {
   let follower: FollowResponse | null = null
   let unfollow: () => void = noop
 
-  const follow$ = new Observable<FollowEventWithRuntime>((observer) => {
-    follower = chainHead(
-      true,
-      (e) => {
-        observer.next(e)
-      },
-      (e) => {
-        follower = null
-        observer.error(e)
-      },
-    )
-    unfollow = () => {
-      observer.complete()
-      follower?.unfollow()
+  const _follow$ = connectable(
+    new Observable<FollowEventWithRuntime>((observer) => {
+      follower = chainHead(
+        true,
+        (e) => {
+          observer.next(e)
+        },
+        (e) => {
+          follower = null
+          observer.error(e)
+        },
+      )
+      unfollow = () => {
+        observer.complete()
+        follower?.unfollow()
+      }
+    }),
+  )
+
+  const startFollow = () => {
+    _follow$.connect()
+    return () => {
+      unfollow()
     }
-  }).pipe(share())
+  }
+
+  const getFollower = () => {
+    if (!follower) throw new Error("Missing chainHead subscription")
+    return follower
+  }
+
+  const getHeader = (hash: string) =>
+    getFollower().header(hash).then(blockHeader.dec)
+
+  const follow$ = _follow$.pipe(
+    withInitializedNumber(getHeader),
+    retryOnStopError(),
+    share(),
+  )
 
   return {
-    getFollower: () => {
-      if (!follower) throw new Error("Missing chainHead subscription")
-      return follower
-    },
-    unfollow: () => {
-      unfollow()
-    },
+    getHeader,
+    getFollower,
+    startFollow,
     follow$,
   }
 }
 
-export const retryOnStopError =
+const retryOnStopError =
   <T extends { type: string }>() =>
   (source$: Observable<T>) =>
     new Observable<
@@ -65,3 +109,7 @@ export const retryOnStopError =
       subscription.add(subscribe())
       return subscription
     })
+
+export type FollowEvent =
+  | ObservedValueOf<ReturnType<ReturnType<typeof withInitializedNumber>>>
+  | { type: "stop-error" }