+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Patch `@polkadot-api/observable-client` connectable bug
+
 ## 0.9.15 - 2024-11-06
 
 ### Fixed