+++ packages/observable-client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Make `follow$` a connectable observable.
+
 ## 0.6.1 - 2024-11-05
 
 ### Fixed